file checking

+++ install.sh
@@ -1,40 +1,50 @@
 #!/bin/bash
 
 echo "---------- Installer script for dotfiles ----------"
-echo "backing up files in home directory"
+echo "asking for sudo access"
+sudo su
 
 # Backups and copying files
 mkdir "backup_dotfiles"
-mv "$HOME/.vimrc" "backup_dotfiles/"
-mv "$HOME/.vim" "backup_dotfiles/"
-mv "$HOME/.config" "backup_dotfiles/.config"
+if (( $? == 0 )); then  # dont overwrite an existing backup
+    echo "backing up current files"
+    mv "$HOME/.vimrc" "backup_dotfiles/"
+    mv "$HOME/.vim" "backup_dotfiles/"
+    mv "$HOME/.config" "backup_dotfiles/.config"
+fi
 
-echo "cloning repositories for vim plugins"
-git clone 'https://github.com/itchyny/lightline.vim.git' "$HOME/.vim/pack/plugins/start/lightline"
-git clone 'https://github.com/frazrepo/vim-rainbow.git' "$HOME/.vim/pack/plugins/start/rainbow"
+echo "checking for vim plugins..."
+if [ ! -e $HOME/.vim/pack/plugins/start/lightline ]; then
+    git clone 'https://github.com/itchyny/lightline.vim.git' "$HOME/.vim/pack/plugins/start/lightline"
 
+if [ ! -e $HOME/.vim/pack/plugins/start/lightline ]; then
+    git clone 'https://github.com/frazrepo/vim-rainbow.git' "$HOME/.vim/pack/plugins/start/rainbow"
+
+# Blocking ads and malware hosts
+echo "blocking ads and malware"
+# https://github.com/StevenBlack/hosts/blob/master/readme.md
+sudo wget -O /etc/hosts https://raw.githubusercontent.com/StevenBlack/hosts/master/hosts
 
 # Installing packages
 echo "installing packages"
-echo "Enter password for sudo: "
 
 if  which apt ; then
     sudo apt install zsh bat exa fd-find fonts-powerline
 elif which pacman ; then
     sudo pacman -Sy --noconfirm --needed firefox zsh bat eza fd git base-devel powerline powerline-fonts i3-wm i3lock i3status-rust nitrogen nautilus
-    echo "installing yet another yogurt"
-    git clone https://aur.archlinux.org/yay-bin.git $HOME/Downloads/yay-bin
-    cd $HOME/Downloads/yay-bin
-    makepkg -si
-    echo "yay installed... installing from AUR"
-    yay -S --noconfirm --needed ttf-meslo-nerd-font-powerlevel10k visual-studio-code-bin spotify
-    yay -Y --gendb
-    yay -Y --devel --diffmenu=false --save
-    cd $HOME/dotfiles
+    if [ ! -e $HOME/Downloads/yay-bin ]; then
+        echo "installing yet another yogurt"
+        git clone https://aur.archlinux.org/yay-bin.git $HOME/Downloads/yay-bin
+        cd $HOME/Downloads/yay-bin
+        makepkg -si
+        echo "yay installed... installing from AUR"
+        yay -S --noconfirm --needed ttf-meslo-nerd-font-powerlevel10k visual-studio-code-bin spotify
+        yay -Y --gendb
+        yay -Y --devel --diffmenu=false --save
+        cd $HOME/dotfiles
+    fi
 else
-    echo "installation failed, make sure apt or pacman is available or edit the install script to work with your package manager"
-    echo "Exiting..."
-    exit
+    echo "installation failed, apt or pacman not available"
 fi
 
 echo "installing oh-my-zsh"
@@ -48,8 +58,8 @@ git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-m
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
 
 echo "copying files"  # copying files after all installs to ensure they aren't overwritten by install processes
+echo "replacing config files"
 cp -r ".zshrc" ".vimrc" ".vim" ".p10k.zsh" ".config" "$HOME"
-
 
 # Configuring shells
 echo 'Changing preferred shell to zsh'
@@ -62,10 +72,9 @@ else
     mkdir -p $HOME/miniconda3
     wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
     bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
-    rm -rf $HOME/miniconda3/miniconda.sh
+    rm $HOME/miniconda3/miniconda.sh
     $HOME/miniconda3/bin/conda init zsh
 fi
-
 
 # Git configuration
 read -n1 -p "Setup git config: [y]/n" configure_git
